+++ CHANGELOG.md
@@ -22,7 +22,10 @@
   This affects all queries whose input was a block or transaction hash.
   These queries now return `InvalidArgument` error, as opposed to `Unknown`
   which they returned previously.
-  
+- Fix issue #244: Collector to keep querying. Remove the parameter for maximum allowed
+  times a gRPC call can fail and keeps `node-collector` querying forever.
+- `GetAccountInfo` endpoint supports querying the account via the account index.
+
 ## concordium-node 3.0.1
 
 - Fix a starvation bug in some cases of parallel node queries.